requirements.txt
gradio>=4.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pyautogui>=0.9.54
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
speechrecognition>=3.10.0
pydub>=0.25.1
pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
opencv-python>=4.8.0
python-dotenv>=1.0.0